* commit '22d78bdfa4d560a0b8a2dbb547574bfe61bf64f0':
  Updates to warnhist script, along with a few cleanups, and add some comments to warning cleanups that are a bit obscure.
  Cleaned up code that genericizes warning output, and add "-u" option to leave unique types in warnings (i.e. don't genericize output).  Also fixed the '-S' option to work correctly.
  Added GCC9 flags to CMake build
  Add support for GCC9, update warnhist script, and clean up warnings.

+++ test/tid.c
@@ -498,7 +498,7 @@ static int test_id_type_list(void)
         goto out;
 
     /* Sanity check */
-    if(startType >= H5I_MAX_NUM_TYPES || startType < H5I_NTYPES)
+    if((int)startType >= H5I_MAX_NUM_TYPES || startType < H5I_NTYPES)
     {
         /* Error condition, throw an error */
         CHECK(1, 1, "H5Iregister_type");
